Refactor to Modules and cleanup
- Refactors to Modules
- Removes default_splits file and uses Run::default()
- General code cleanup
