+++ src/bin/chronos-splits-editor/global_state.rs
@@ -1,0 +1,7 @@
+use livesplit_core::run::Editor;
+
+pub struct GlobalState {
+    pub splits_editor: Editor,
+    pub splits_file: String,
+}
+
