+++ src/bin/chronos-splits-editor/views/mod.rs
@@ -1,0 +1,3 @@
+pub mod layout;
+pub mod splits;
+pub mod details;
