--- src/default_splits.lss
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><Run version="1.8.0"><GameIcon/><GameName/><CategoryName/><Metadata><Run id=""/><Platform usesEmulator="False"/><Region/><SpeedrunComVariables/><CustomVariables/></Metadata><Offset>00:00:00.000000000</Offset><AttemptCount>0</AttemptCount><AttemptHistory/><Segments><Segment><Name>Time</Name><Icon/><SplitTimes><SplitTime name="Personal Best"/></SplitTimes><BestSegmentTime/><SegmentHistory/></Segment></Segments><AutoSplitterSettings/></Run>
